+++ equinox-web-csharp/Web/EventStoreContext.cs
@@ -51,8 +51,8 @@ namespace TodoBackendTemplate
             return new Context(conn, new BatchingPolicy(maxBatchSize: 500));
         }
 
-        public override Func<Target, IStream<TEvent, TState>> Resolve<TEvent, TState>(
-            FsCodec.IUnionEncoder<TEvent, byte[], object> codec,
+        public override Func<string, IStream<TEvent, TState>> Resolve<TEvent, TState>(
+            FsCodec.IEventCodec<TEvent, byte[], object> codec,
             Func<TState, IEnumerable<TEvent>, TState> fold,
             TState initial,
             Func<TEvent, bool> isOrigin = null,