+++ equinox-web-csharp/Web/MemoryStoreContext.cs
@@ -15,8 +15,8 @@ namespace TodoBackendTemplate
         public MemoryStoreContext(VolatileStore<object> store) =>
             _store = store;
 
-        public override Func<Target, IStream<TEvent, TState>> Resolve<TEvent, TState>(
-            FsCodec.IUnionEncoder<TEvent, byte[], object> codec,
+        public override Func<string, IStream<TEvent, TState>> Resolve<TEvent, TState>(
+            FsCodec.IEventCodec<TEvent, byte[], object> codec,
             Func<TState, IEnumerable<TEvent>, TState> fold,
             TState initial,
             Func<TEvent, bool> isOrigin = null,