+++ propulsion-summary-consumer/Infrastructure.fs
@@ -6,28 +6,14 @@ open System
 module EventCodec =
 
     /// Uses the supplied codec to decode the supplied event record `x` (iff at LogEventLevel.Debug, detail fails to `log` citing the `stream` and content)
-    let tryDecode (codec : FsCodec.IUnionEncoder<_, _, _>) (log : Serilog.ILogger) (stream : string) (x : FsCodec.ITimelineEvent<byte[]>) =
+    let tryDecode (codec : FsCodec.IEventCodec<_, _, _>) (log : Serilog.ILogger) streamName (x : FsCodec.ITimelineEvent<byte[]>) =
         match codec.TryDecode x with
         | None ->
             if log.IsEnabled Serilog.Events.LogEventLevel.Debug then
                 log.ForContext("event", System.Text.Encoding.UTF8.GetString(x.Data), true)
-                    .Debug("Codec {type} Could not decode {eventType} in {stream}", codec.GetType().FullName, x.EventType, stream)
+                    .Debug("Codec {type} Could not decode {eventType} in {stream}", codec.GetType().FullName, x.EventType, streamName)
             None
         | x -> x
-
-[<AutoOpen>]
-module StreamName =
-    let private catSeparators = [|'-'|]
-    let private split (sep : char[]) (streamName : string) = streamName.Split(sep, 2, StringSplitOptions.RemoveEmptyEntries)
-    let category (streamName : string) = let fragments = split catSeparators streamName in fragments.[0]
-    let (|Two|_|) (sep : char[]) value =
-        match split sep value with
-        | [| category; id |] -> Some (category, id)
-        | _ -> None
-    let (|Category|Other|) (streamName : string) =
-        match streamName with
-        | Two catSeparators (category, id) -> Category (category, id)
-        | _ -> Other streamName
 
 module Guid =
     let inline toStringN (x : Guid) = x.ToString "N"
@@ -38,3 +24,4 @@ and [<Measure>] clientId
 module ClientId =
     let toString (value : ClientId) : string = Guid.toStringN %value
     let parse (value : string) : ClientId = let raw = Guid.Parse value in % raw
+    let (|Parse|) = parse